removed mailer files. No more new routes, only signup routes for user and admin. Created Error Helper to manage errors. Application helper added for isloggedin and current user. admin checker added to user model, and attached to application in error helper. basic nav bar added. factions model updated to include an info section

+++ app/controllers/users_controller.rb
@@ -1,6 +1,8 @@
 class UsersController < ApplicationController
     def show
-        @user = User.find(params[:id])
+        if (logged_in? && current_user.id === session[:user_id]) || admin?
+            @user = User.find(params[:id])
+        end
     end 
 
     def new 
@@ -8,40 +10,59 @@ class UsersController < ApplicationController
     end 
 
     def create
-        if user_params[:code] != ""
-            if User.is_admin?(user_params[:code])
+        if user_params[:code] != nil
+            if User.find_by(email: user_params[:email])
+                email_taken
+                redirect_to admin_signup_path
+            else
+                if admin?(user_params[:code])
+                    user = User.new(user_params)
+                    if user.save
+                        session[:user_id] = user.id
+                        session[:code] = user.code
+                        redirect_to admin_factions_path
+                    end
+                else
+                    not_admin
+                    redirect_to admin_signup_path
+                end
+            end
+        else 
+            if User.find_by(email: user_params[:email])
+                email_taken
+                redirect_to signup_path
+            else
                 user = User.new(user_params)
                 if user.save
                     session[:user_id] = user.id
-                    redirect_to admin_user_path(user)
+                    redirect_to user_path(user)
+                else
+                    save_issue
+                    redirect_to new_user_path
                 end
-            else 
-                redirect_to new_admin_user_path
-            end
-        else 
-            user = User.new(user_params)
-            if user.save
-                session[:user_id] = user.id
-                redirect_to user_path(user)
-            else 
-                redirect_to user_new_path
             end
         end
     end
 
-    def edit 
-        @user = User.find(params[:id])
+    def edit
+        if (logged_in? && current_user.id === session[:user_id]) || admin?
+            @user = User.find(params[:id])
+        end
     end 
 
-    def update 
-        user = User.find(params[:id])
-        user.update(user_params)
-        redirect_to user_path(user)
+    def update
+        if (logged_in? && current_user.id === session[:user_id]) || admin?
+            user = User.find(params[:id])
+            user.update(user_params)
+            redirect_to user_path(user)
+        end
     end
 
     def destroy
-        user = User.find(params[:id])
-        user.destroy
+        if (logged_in? && current_user.id === session[:user_id]) || admin?
+            user = User.find(params[:id])
+            user.destroy
+        end
     end
 
     private
